Rename the ObjectPrinter tool to FbxPrint.
